rename checkbox and radiobutton components
